@@ -16,15 +16,13 @@
 
 package com.faendir.rhino_android;
 
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 import com.android.dex.Dex;
+import dalvik.system.PathClassLoader;
 
 import java.io.File;
 import java.io.IOException;
-
-import dalvik.system.PathClassLoader;
 
 /**
  * @author F43nd1r
