@@ -17,16 +17,13 @@
 package com.faendir.rhino_android;
 
 import android.os.Build;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.annotation.RequiresApi;
-
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.annotation.RequiresApi;
 import com.android.dex.Dex;
-
-import java.io.IOException;
-import java.nio.ByteBuffer;
-
 import dalvik.system.InMemoryDexClassLoader;
+
+import java.nio.ByteBuffer;
 
 /**
  * @author F43nd1r
